+++ docs/examples/UserGuide/misc_tips.jl
@@ -24,4 +24,10 @@ sqc(qry) = @chain t(qry) begin
             end;
 
 # Call this function at the end of a chain similar the `@show_query` or`@collect` macros
+# _printed query is not seen here as it prints to the REPL_
 @chain t(dfv) @summarize() sqc()
+
+# ## Color Printing
+# Queries print with some code words in color to the REPL. To turn off this feature run the one of the following.
+#   - `TidierDB.color[] = false`
+#   - `DB.color = false`